feat: add update post and move actions to repository

+++ src/actions/post/delete-post-action.ts
@@ -4,7 +4,6 @@ import { drizzleDb } from "@/db/drizzle";
 import { postsTable } from "@/db/drizzle/schemas";
 import { asyncDelay } from "@/lib/async-delay";
 import { postRepository } from "@/repositories/post";
-
 import { logColor } from "@/utils/log-color";
 import { eq } from "drizzle-orm";
 import { revalidateTag } from "next/cache";
@@ -22,22 +21,31 @@ export async function deletePostAction(id: string) {
     };
   }
 
-  const post = await postRepository.findById(id).catch(() => undefined);
+  let post;
+  try {
+    post = await postRepository.delete(id);
+  } catch (e: unknown) {
+    if (e instanceof Error) {
+      return {
+        error: e.message,
+      };
+    }
 
-  if (!post) {
+    if (!post) {
+      return {
+        error: "Erro desconhecido",
+      };
+    }
+
+    // TODO: mover este método para o repositório
+    await drizzleDb.delete(postsTable).where(eq(postsTable.id, id));
+
+    // TODO: revalidateTag ou revalidatePath
+    revalidateTag("posts");
+    revalidateTag(`post-${post.slug}`);
+
     return {
-      error: "Post não existe",
+      error: "",
     };
   }
-
-  // TODO: mover este método para o repositório
-  await drizzleDb.delete(postsTable).where(eq(postsTable.id, id));
-
-  // TODO: revalidateTag ou revalidatePath
-  revalidateTag("posts");
-  revalidateTag(`post-${post.slug}`);
-
-  return {
-    error: "",
-  };
 }
